refactor: fusion page protection des données avec mentions légales
