+++ ui/modules/dashboard/DashboardOrganisme.tsx
@@ -688,7 +688,7 @@ const DashboardOrganisme = ({ organisme, modePublique }: Props) => {
                       données peuvent être consultées <strong>exclusivement</strong> par votre organisme et les
                       administrations publiques dans le cadre de la{" "}
                       <Link
-                        href="/protection-des-donnees"
+                        href="/mentions-legales#protection-des-donnees"
                         borderBottom="1px"
                         _hover={{ textDecoration: "none" }}
                         display="inline-flex"
